ci: Consolidate build stage
The build stage does not produce any artifact needed for the test stage.
So we can literally just move the build job to the test stage in order
to parallelize efforts.

The update_snapshots job is moved to the pre-build stage as it is a
manual job anyway.

All jobs that do not depend on the results of the build job now have
this dependency removed.