fix(storybook): transpile object-spread for Edge
Edge errors when loading storybook due to object-spreads used
within bootstrap-vue and some stories.

This commit adds a webpack-loader to handle transpilation for all
modules that use object-spread.

Note: The webpack-loader is only added to the list of rules if
NODE_ENV is not 'test'.

This is because the settings in 'babel.conf' target 'node' for tests,
which is needed for jest but breaks the rendering of storyshots if
the loader is applied.