ci: Disable verbose webpack logging on CI only
The ProgressPlugin generates a lot of unnecessary lines in CI and the
resulting build logs are half a megabyte large. While having the
progress displayes locally might be valueable, we can save a bit on
storage and make it easier to analyse when something is wrong.

Webpack should still output errors in case compilation fails.