revert: fix(Link): Add support to allow only safe urls
This reverts merge request
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/1362

The changes introduced in the GlLink component caused some failures in
GitLab, preventing us from upgrading the `@gitlab/ui` package. We are
going to revert those changes for now until we can bring them back in
along with an integration MR to make it easier to upgrade the package
next time they are published.