Address linting issues
Overrides have been added to ESLint config to ignore the files that
violate the import/no-default-export rule. Some of these overrides are
permanent, as it makes sense for some files to export default modules.
Temporary otherwise are marked as such in the config and will have to be
migrated to named exports in follow-up issues.

For some files that don't follow any particular naming pattern but where
it makes sense to keep a default export, the rule has been disabled via
a eslint-disable comment.

Additionally, this removes eslint-disable comments for the
import/prefer-default-export rule, which isn't used anymore.