chore(deps): update @gitlab/eslint-plugin
This updates to @gitlab/eslint-plugin@6, which includes breaking changes
via the transitive dependency eslint-plugin-vue being upgraded from v6
to v7.

This adds/improves various lint rules of Vue files; the full changes are
documented in their [release post][1]. The most notable addition is the
vue/no-mutating-props rule, which can help to prevent subtle runtime
bugs.

Trivial violations of `vue/no-lone-template` and
`vue/order-in-components` are fixed in this change, but there are two
violations of `vue/no-mutating-props` that are not trivially fixed, so
the rule has been disabled for those to areas of code.

Addresses https://gitlab.com/gitlab-org/gitlab-ui/-/issues/1113, part of
https://gitlab.com/groups/gitlab-org/-/epics/5142.

[1]: https://github.com/vuejs/eslint-plugin-vue/releases/tag/v7.0.0