chore(jest): Upgrade jest
Upgrade jest to the latest version. It also took me a few seconds to
find our jest config, so moving it from the package.json to a separate
file.

The only thing that changed, tests which are checking the focus of
elements need to be attached to the DOM. There is also better support
for CSS custom properties and how they are queried.