Add stack-rendered sprite to dist
It would make sense to switch to stacked rendering because it allows us
to re-use the sprite icon in CSS, but we had issues with it in the past.

By adding a separate dist icon sprite we are able to switch and test
everything properly rather than breaking things.

See: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/2881