fix: Work around vue-apollo@3 bug
While GitLab UI doesn't directly use vue-apollo@3, the `FilteredSearch`
component can generate errors in projects which do (like GitLab).

The problem is that under certain circumstances, vue-apollo [invokes][1]
a component's `provide` option method during the its `beforeCreate`
lifecycle event. This is too early, and breaks the [expectation][2] that
props can be referenced in a component's `provide` option method.

The workaround here is to return a function that returns the component's
prop value, rather than the plain value itself. This works because the
prop is read lazily, when the descendant component calls the provided
function, by which time the ancestor's props will have been initialised.

[1]: https://github.com/vuejs/vue-apollo/blob/v3.0.3/packages/vue-apollo/src/mixin.js#L17-L25
[2]: https://github.com/vuejs/vue/blob/v2.6.14/src/core/instance/init.js#L58

Co-authored-by: Mark Florian <mflorian@gitlab.com>