ci(forks): ensure that fork pipelines are executed correctly
We were comparing against CI_MERGE_REQUEST_SOURCE_PROJECT_PATH which
always points to the fork instead of the project the pipeline is run in.

So for example we have an MR created from fork/gitlab-ui and the person
who forked gitlab-ui created the MR, the pipeline was executed in the
context of: `fork/gitlab-ui`. If then a GitLab team member ran the `Run
pipeline` in the context of the MR, the pipeline was executed in context
of `gitlab-org/gitlab-ui`, but some of the jobs, e.g. `review`,
`create_integration_branch` were not executed.

By utilizing CI_PROJECT_PATH we fix this, because:

1. Pipelines run in `fork/gitlab-ui` will point to that
2. Pipelines run in `gitlab-org/gitlab-ui` will point to that