refactor(GlAlert): isolate limited width styles
Separates limited layout width styles from the rest of the rules. This
is made to provide more flexibility when styling HAML alerts in the
product as we'll be able to apply gl-alert-max-content styles to
properly position the elements, combined with GitLab classes to set the
alert content's width so that it matches the main content's width.