fix(Avatar): Add support for multiple-character emojis
Currently the `getAvatarChar` helper only returns the first emoji if the
emoji is actually composed of multiple emoji (e.g. pirate flag, emoji
with gender / skin-color modifiers). This fixes that.

Furthermore we got a report in
https://gitlab.com/gitlab-org/gitlab/-/issues/349420 that older browsers
break with the Unicode property escapes [ref].

We already use the npm-module `emoji-regex` in GitLab to detect emoji
and it is backwards compatible with those older browsers.

[ref]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions/Unicode_Property_Escapes