Don't run publish-to-npm-dry-run on MRs from forks
Prior to this, the job would fail with something like:

    [semantic-release] › ✖  ERELEASEBRANCHES The release branches are
    invalid in the `branches` configuration.  A minimum of 1 and
    a maximum of 3 release branches are required in the branches
    configuration
    (https://github.com/semantic-release/semantic-release/blob/master/docs/usage/configuration.md#branches).
    This may occur if your repository does not have a release branch,
    such as master.
    Your configuration for the problematic branches is [].

This is because the `CI_MERGE_REQUEST_SOURCE_BRANCH_NAME` points to the
branch name on the fork repository. But, semantic-release checks the
repository from `package.json`'s `repository` URL for the existence of
that branch, so it fails.

See also
https://github.com/semantic-release/semantic-release/issues/1645 and
https://github.com/semantic-release/semantic-release/issues/1985#issuecomment-1024977423.