fix(GlFilteredSearch): Keep active state on backspace on first token
This is a follow-up to
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/2705, which
changed how the active state of tokens is managed on token destruction.

It introduced a small usability issue where if you repeatedly backspace
to clear `GlFilteredSearch` of all tokens, you could no longer enter
_new_ tokens (though you could enter text), and you had to blur and then
focus the component to fix it.

This change makes sure that if you backspace-to-delete the first token,
then the _new_ first token will be activated.