docs(GlSkeletonLoader): Improve story controls and layout
The `width` and `height` props are a bit confusing on this component,
and they _appear_ to have differing effects between its two stories.

The `width` and `height` props modify the SVG's `viewBox` attribute,
which defines the position and dimension of the SVG viewport. This means
that shapes in the SVG that use absolute values for their size and
position get scaled in the opposite direction to what you'd expect,
whereas sizes and positions defined with percentages scale with the
`width` and `height`, as you'd expect.

In the default story, where no slot content/custom shapes are provided,
the component generates shapes using percentages for sizing. In the
custom shapes story, the shapes are defined using absolute values. This
is why those controls *appear* to have differing behaviour between the
two stories.