test: prevent listbox from opening in e2e tests
This adds a control to prevent the listbox from opening on mount when
running integration tests against it. This ensures that the dropdown
menu is not visible to satisfy the first assertion.
For some reason, this worked previously without this tweak. The fact
that we need to make this change now might be due to the fact that we've
enabled bundle splitting, which lets stories' code execute faster than
before.