ci: Lint against $slots usage
Just ignore existing violations for now, since there are some subtleties
when migrating from the $slots API to $scopedSlots. These will have to
be migrated separately later.

The main subtlety is that these two expressions aren't equivalent:

    Boolean(this.$slots.foo)

    Boolean(this.$scopedSlots.foo)

The former evaluates to `false` if either the `foo` slot isn't provided,
or if the provided `foo` slot has no content!

The latter only evaluates to `false` if the `foo` slot is not provided.

The simplest way to make these equivalent would be to call the slot
function in the second case:

    Boolean(this.$scopedSlots.foo())

But, depending on where this code lives, it may be called too often, or
worse, have side effects.

The best way to migrate each of these will need to be worked out
case-by-case.

Part of https://gitlab.com/groups/gitlab-org/-/epics/8367.