fix: convert gl-sm-pr-2 to mobile-first, remove gl-md-pr-2
This converts the gl-sm-pr-2 to mobile-first and removes gl-md-pr-2
which won't be used anymore once those changes are integrated into
GitLab.

BREAKING CHANGE: This converts gl-sm-pr-2 from desktop-first to
mobile-first and removes gl-md-pr-2. If you were relying on the former,
please migrate to a mobile-first approach. If relying on the latter, you
might need to add it back to the library. We are defaulting to removing
utils that aren't used in GitLab projects while we remove desktop-first
utils, so it might be necessary to re-add some of them if they are used
elsewhere.