chore(GlListbox): update group example
This complexifies the Group story to have selected items move to
dedicated groups thanks to exposing the items as a computed prop.
This lets us demonstrate a bug where the listbox would unexpectedly
close when selecting an item that immediately gets removed from the DOM.
The bug is due to our `outside` directive that doesn't capture click
events and thus considers that the click happened outside of the
listbox.