ci: Ensure visual job is run in merge trains
The visual job wasn't running in merge train pipelines because the
CI_MERGE_REQUEST_IID check always evaluates to true on merge trains,
preventing the merge train check from being evaluated.

This made the the visual job manual in those cases, instead of running
normally.

In pseudocode, we were doing this:

    if branch name is 'main' and project is 'gitlab-org/gitlab-ui':
        run visual job
    else if is merge request pipeline:
        set visual job to manual
    else if is merge train pipeline:
        run visual job

The fix is to reverse the order of the last two checks, so that the more
specific one (is it a merge train) is done before the less specific one
(is it a merge request pipeline).

This also adds the missing snapshots that were breaking the main branch.