feat(GlDisclosureDropdown): Loosen action item type
This allows for a simpler way of specifying action items. They no longer
need an `action` method. They simply must not have an `href` string.

This means that if all items are actions, you can simply provide their
text, and then add an `action` event listener to the root component,
rather than add an action listener to each item. It is also still
possible to add unique `action` handlers for each item.

BREAKING CHANGE: Action items are now determined by the *lack* of an
href string. An item with an href string is rendered as a link item,
whether or not it has an action callback.

Before this change, if an item had *both* an href and an action, it
would render as a button. Now, it would render as an anchor.