fix(GlFormDate): fix feedback text size
It was being rendered too small, according to Pajamas. This was due to
Bootstrap 4.5.3's `invalid-feedback` class, which sets `font-size: 80%`
(computing to 12.8px).

Bootstrap 4.6.2 includes https://github.com/twbs/bootstrap/pull/23816,
which changes `invalid-feedback` such that the font size computes to the
value we want: 14px, resulting in this storyshot change.

This does not affect feedback text rendered by GlFormGroup, since that
component's styles override the font size:
https://gitlab.com/gitlab-org/gitlab-ui/blob/v56.2.0/src/components/base/form/form_group/form_group.scss#L24

It's unusual that GlFormDate renders its own invalid-state feedback, as
other form components rely on GlFormGroup for that. This was discussed
in the original implementation merge request:
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/3115#note_1152044321.