refactor(GlFilteredSearch): Improve suggestion selection logic
This rewrites the suggestion selection tracking logic, and adds two test
cases:

- One is a regression test, as it fails on the previous implementation.
- The other just increases test coverage for the new implementation.

This new implementation seems a bit over-engineered at first, but isn't
really, since it makes it easier to switch to purely cyclic selection in
future. That is, not allowing deselection of options via the arrow keys.
This will be needed for the upcoming work to stop treating spaces
specially.

See https://gitlab.com/gitlab-org/gitlab-ui/-/issues/2159 and
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/3226 for more
detail.