test(GlPopover): Reduce test flakiness
It seems that @vue/compat ends up destroying the popover element faster
than Vue 2, making a race more likely.

Why's it a race? The problem is the use of the `findByTestId` command.

In this popover test, if the popover element is destroyed fast enough,
we're doing something like:

    cy.get('i-do-not-exist').first().should('not.exist');

That will always fail, since there's no "first" element to assert
should not exist.

We simply need to remove the `.first()` call to fix this.